+++ lib/matrix.dart
@@ -74,6 +74,90 @@ abstract class Matrix  implements Iterable<Iterable<double>> {
     }
   }
 
+  /// Creates a matrix, where elements from [source] are the elements for the
+  /// matrix main diagonal, the rest of the elements are zero
+  ///
+  /// ````dart
+  /// final matrix = Matrix.diagonal([1, 2, 3, 4, 5]);
+  ///
+  /// print(matrix);
+  /// ````
+  ///
+  /// The output:
+  ///
+  /// ```
+  /// Matrix 5 x 5:
+  /// (1.0, 0.0, 0.0, 0.0, 0.0)
+  /// (0.0, 2.0, 0.0, 0.0, 0.0)
+  /// (0.0, 0.0, 3.0, 0.0, 0.0)
+  /// (0.0, 0.0, 0.0, 4.0, 0.0)
+  /// (0.0, 0.0, 0.0, 0.0, 5.0)
+  /// ```
+  factory Matrix.diagonal(List<double> source, {DType dtype = DType.float32}) {
+    switch (dtype) {
+      case DType.float32:
+        return Float32Matrix.diagonal(source);
+      default:
+        throw UnimplementedError('Matrix of type $dtype is not implemented yet');
+    }
+  }
+
+  /// Creates a matrix of [size] * [size] dimension, where all the main
+  /// diagonal elements are equal to [scalar], the rest of the elements are 0
+  ///
+  /// ````dart
+  /// final matrix = Matrix.scalar(3, 5);
+  ///
+  /// print(matrix);
+  /// ````
+  ///
+  /// The output:
+  ///
+  /// ```
+  /// Matrix 5 x 5:
+  /// (3.0, 0.0, 0.0, 0.0, 0.0)
+  /// (0.0, 3.0, 0.0, 0.0, 0.0)
+  /// (0.0, 0.0, 3.0, 0.0, 0.0)
+  /// (0.0, 0.0, 0.0, 3.0, 0.0)
+  /// (0.0, 0.0, 0.0, 0.0, 3.0)
+  /// ```
+  factory Matrix.scalar(double scalar, int size, {DType dtype = DType.float32}) {
+    switch (dtype) {
+      case DType.float32:
+        return Float32Matrix.scalar(scalar, size);
+      default:
+        throw UnimplementedError('Matrix of type $dtype is not implemented yet');
+    }
+  }
+
+  /// Creates a matrix of [size] * [size] dimension, where all the main
+  /// diagonal elements are equal to 1, the rest of the elements are 0
+  ///
+  /// ````dart
+  /// final matrix = Matrix.identity(5);
+  ///
+  /// print(matrix);
+  /// ````
+  ///
+  /// The output:
+  ///
+  /// ```
+  /// Matrix 5 x 5:
+  /// (1.0, 0.0, 0.0, 0.0, 0.0)
+  /// (0.0, 1.0, 0.0, 0.0, 0.0)
+  /// (0.0, 0.0, 1.0, 0.0, 0.0)
+  /// (0.0, 0.0, 0.0, 1.0, 0.0)
+  /// (0.0, 0.0, 0.0, 0.0, 1.0)
+  /// ```
+  factory Matrix.identity(int size, {DType dtype = DType.float32}) {
+    switch (dtype) {
+      case DType.float32:
+        return Float32Matrix.scalar(1.0, size);
+      default:
+        throw UnimplementedError('Matrix of type $dtype is not implemented yet');
+    }
+  }
+
   /// A data type of [Matrix] elements
   DType get dtype;